+++ include/multiband_compressor.hpp
@@ -42,7 +42,9 @@ class MultibandCompressor : public PluginBase {
   void process(std::span<float>& left_in,
                std::span<float>& right_in,
                std::span<float>& left_out,
-               std::span<float>& right_out) override;
+               std::span<float>& right_out,
+               std::span<float>& probe_left,
+               std::span<float>& probe_right) override;
 
   sigc::signal<void(const float&)> latency;
 
@@ -59,6 +61,8 @@ class MultibandCompressor : public PluginBase {
   uint latency_n_frames = 0U;
 
   std::unique_ptr<lv2::Lv2Wrapper> lv2_wrapper;
+
+  std::vector<pw_proxy*> list_proxies;
 };
 
 #endif